refactor: 인증 상세 인증 컴포넌트에서 progressBar 컴포넌트로 변경
